+++ README.md
@@ -4,10 +4,11 @@
 
 - [x] 딜러와 플레이어 중 카드의 합이 21 or 21에 가장 가까운 숫자를 가지는쪽이 이긴다.
 - [x] 게임이 시작되면 플레이어는 2장의 카드를 지급 받는다.
+  - [x] 최초 받은 패의 값이 21이면 blackjack이며 게임은 종료된다.
 - 유저
   - [x] 유저는 카드를 뽑을 수 있는지 여부를 판단할 수 있다.
-  - [x] 유저의 패의 합이 21 미만일 경우 유저는 카드를 더 받을 수 있거나 거절할 수 있다. 
-  - [x] 유저의 패의 합이 21 이상일 경우 유저는 카드를 더 받을 수 없다. 
+  - [x] 유저의 패의 합이 21 미만일 경우 유저는 카드를 draw 하거나 stay 할 수 있다. 
+  - [x] 뽑았을때 유저의 패의 합이 21 이상이 되면 유저는 bust 되며 게임은 종료된다. 
 - 카드
   - [x] 카드는 카드번호와 마크로 구성된다. 
   - [x] 카드는 Ace 여부를 판단할 수 있다.
@@ -17,4 +18,3 @@
 - 덱
   - [x] 하나의 덱은 카드의 모음으로 이루어진다.
   - [x] 덱은 하나의 카드를 무작위로 뽑을 수 있다.
-
