+++ src/main/kotlin/blackjack/domain/CardMark.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class CardMark {
     SPADE,
