+++ src/main/kotlin/blackjack/domain/Deck.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 fun interface Deck {
     fun draw(): Card?
