+++ src/test/kotlin/blackjack/DeckTest.kt
@@ -1,5 +1,7 @@
 package blackjack
 
+import blackjack.domain.Card
+import blackjack.infrastructure.RandomDeck
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.types.shouldBeInstanceOf
